@@ -1,16 +1,32 @@
-import 'package:GTUBT/ui/blocs/page_bloc/bloc.dart';
 import 'package:equatable/equatable.dart';
 import 'package:flutter/widgets.dart';
 
 abstract class PageEvent extends Equatable {
-  PageEvent([List props = const []]) : super(props);
+  PageEvent() : super();
 }
 
-class PageChanged extends PageEvent {
+class NavBarPageChanged extends PageEvent {
   final int page;
-  
-  PageChanged({@required this.page}) : super([page]);
+  final BuildContext context;
+
+  NavBarPageChanged({@required this.page, @required this.context}) : super();
 
   @override
   String toString() => 'PageChanged { Page :$page }';
+
+  @override
+  List<Object> get props => [page, context];
+}
+
+class PageChanged extends PageEvent {
+  final BuildContext context;
+  final String routeName;
+
+  PageChanged({@required this.context, @required this.routeName});
+
+  @override
+  String toString() => 'PageChanged {routeName: $routeName}';
+
+  @override
+  List<Object> get props => [context, routeName];
 }
