@@ -1,33 +1,44 @@
+import 'package:GTUBT/service/authentication/authentication.dart';
+import 'package:GTUBT/ui/blocs/authentication_bloc/bloc.dart';
+import 'package:GTUBT/ui/blocs/register_bloc/bloc.dart';
 import 'package:GTUBT/ui/blocs/user_bloc/bloc.dart';
-import 'package:flutter/material.dart';
 import 'package:bloc/bloc.dart';
+import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
-import 'ui/blocs/page_bloc/bloc.dart';
-import 'ui/routes.dart';
 import 'package:onesignal_flutter/onesignal_flutter.dart';
 
-void main()  {
+import 'ui/blocs/page_bloc/bloc.dart';
+import 'ui/routes.dart';
+
+void main() {
   BlocSupervisor.delegate = SimpleBlocDelegate();
   runApp(MyApp());
 }
 
 class MyApp extends StatelessWidget {
-  // This widget is the root of your application.
-
   @override
   Widget build(BuildContext context) {
-    OneSignal.shared.init(
-        "3f4bf8af-bd07-4353-bdb9-ed92a96908aa",
-        iOSSettings: {
-          OSiOSSettings.autoPrompt: false,
-          OSiOSSettings.inAppLaunchUrl: true
-        }
-    );
-    OneSignal.shared.setInFocusDisplayType(OSNotificationDisplayType.notification);
+    final AuthService _authService = AuthService();
+    OneSignal.shared.init("3f4bf8af-bd07-4353-bdb9-ed92a96908aa", iOSSettings: {
+      OSiOSSettings.autoPrompt: false,
+      OSiOSSettings.inAppLaunchUrl: true
+    });
+    OneSignal.shared
+        .setInFocusDisplayType(OSNotificationDisplayType.notification);
     return MultiBlocProvider(
       providers: [
-        BlocProvider<PageBloc>(create: (context) => PageBloc(),),
-        BlocProvider<UserBloc>(create: (context) => UserBloc(),),
+        BlocProvider<AuthenticationBloc>(
+          create: (context) => AuthenticationBloc(authService: _authService),
+        ),
+        BlocProvider<PageBloc>(
+          create: (context) => PageBloc(),
+        ),
+        BlocProvider<UserBloc>(
+          create: (context) => UserBloc(),
+        ),
+        BlocProvider<RegisterBloc>(
+          create: (context) => RegisterBloc(authService: _authService),
+        )
       ],
       child: MaterialApp(
         title: 'GTU BT',
@@ -39,7 +50,6 @@ class MyApp extends StatelessWidget {
         initialRoute: '/',
       ),
     );
-
   }
 }
 
