@@ -1,4 +1,3 @@
-import 'package:GTUBT/models/user.dart';
 import 'package:bloc/bloc.dart';
 import 'user_event.dart';
 import 'user_state.dart';
@@ -7,7 +6,7 @@ class UserBloc extends Bloc<UserEvent, UserState> {
   @override
   get initialState => UserState(user: null);
 
-  UserState currentState;
+  UserState currentState = UserState.loggedOut();
 
   @override
   Stream<UserState> mapEventToState(event) async* {
