@@ -2,12 +2,17 @@ import 'package:GTUBT/models/user.dart';
 import 'package:flutter/foundation.dart';
 
 class UserState {
-  final User user;
-  final bool editMode;
+  User user;
+  bool editMode;
+
   UserState({
     @required this.user,
     this.editMode=false,
   });
+
+  factory UserState.loggedOut() {
+    return UserState(user: null, editMode: false);
+  }
 
   @override
   String toString() {
