@@ -1,18 +1,20 @@
+import 'package:GTUBT/ui/pages/calendar_page.dart';
+import 'package:GTUBT/ui/pages/home_page.dart';
 import 'package:GTUBT/ui/pages/login_page.dart';
+import 'package:GTUBT/ui/pages/main_page.dart';
+import 'package:GTUBT/ui/pages/profile_page.dart';
+import 'package:GTUBT/ui/pages/signup_page.dart';
 import 'package:flutter/material.dart';
-import 'pages/main_page.dart';
-import 'pages/home_page.dart';
-import 'pages/calendar_page.dart';
-import 'pages/profile_page.dart';
-
 
 const ROOT_URL = '/';
 const LOGIN_URL = '/login';
+const SIGN_UP_URL = '/sign_up';
 
 class Routes {
   static final routes = <String, WidgetBuilder>{
-    ROOT_URL: (context) => ExamplePage(),
+    ROOT_URL: (context) => MainPage(),
     LOGIN_URL: (context) => LoginPage(),
+    SIGN_UP_URL: (context) => SignUpPage(),
   };
 
   static final bodyTitle = ["Home", "Calendar", "Profile"];
