@@ -9,5 +9,7 @@ class ColorSets {
   static const defaultBarBackgroundColor = Color(0xFFFFFFFF);
   static const defaultTextColor = Colors.black;
 
-  static const pageBackgroundColor = Color (0xFFFFFFFF);
+  static const pageBackgroundColor = Color(0xFFFFFFFF);
+
+  static const snackBarErrorColor = Color(0xFFF44336);
 }
